adding mocked requests to get some data and figure out how to parse it
